collision are now handled !

+++ src/main/kotlin/model/map/Chunk.kt
@@ -2,7 +2,7 @@ package model.map
 
 
 data class Chunk(
-    val length: Float,
+    val length: Double,
     val platforms: Set<Platform>,
-    val movementFacility: Float = 1f
+    val movementFacility: Double = 1.0
 )
